requirements-dev.txt
pre-commit>=3.7.0
ruff>=0.9.0
lxml>=5.1.0
pymarkdownlnt>=0.9.16
yamllint>=1.35.1
pillow==12.1.1
requests>=2.31.0
